zigbee-switch: Fix color handling for Ikea RGB bulbs
The original code wasn't handling the setHue and setSaturation commands
correctly. The light would flash but not change to the desired color. To
fix the problem, a short timer is started when receiving either the setHue
or setSaturation commands instead of sending a command to the device right
away. The hue (or saturation) value is cached temporarily, and if the other
command is received before the timer expires then both the cached value and
the value from the second command are used to calculate the x/y and are sent
to the device. If only one of the two commands is sent before the timer
expires then the value from that command is combined with the latest known
color state to calculate the x/y.

https://smartthings.atlassian.net/browse/CHAD-8948